// src/app/modules/hcm/shared/models/masters/payroll-calendar-detail.ts
export class PayrollCalendarDetail {
    id: number = -1;
    idPayrollTypeDetail: number = -1;
    payrollType: string = "";
    idCurrency: number= -1;
    currency: string = "";
    year: number= -1;
    numMonday: number = -1;
    period: number = -1;
    conversionFactor: number = -1;
    indMultiply: boolean = true;
    startDate: string = "";
    finishDate: string = "";
    paymentDate: string = "";
    idStatus: number = -1;
    status: string = "";
    paymentDay: number = 0;
    userCreate: string = "";
    userUpdate: string = ""; 
}